+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - [Issue #736](https://github.com/influxdb/influxdb/issues/736). Only db admins should be able to drop a series
 - [Issue #713](https://github.com/influxdb/influxdb/issues/713). Null should be a valid fill value
 - [Issue #644](https://github.com/influxdb/influxdb/issues/644). Graphite api should write data in batches to the coordinator
+- [Issue #740](https://github.com/influxdb/influxdb/issues/740). Panic when distinct fields are selected from an inner join
+- [Issue #781](https://github.com/influxdb/influxdb/issues/781). Panic when distinct fields are added after an inner join
 
 ## v0.8.0-rc.3 [2014-07-21]